Rebase / Merge (#2)
* Add possibility to run once 

In some instances (e.g. external scheduler, orchestration solutions), it might be desirable to run only once and let a container-external solution do the scheduling.

With this patch, such a behavior can be achieved by setting ``RUN_ONCE=1`` as a new config option of the container.

* Update entrypoint

* Add possibility to run once

In some instances (e.g. external scheduler, orchestration solutions), it might be desirable to run only once and let a container-external solution do the scheduling.

With this patch, such a behavior can be achieved by setting ``RUN_ONCE=1`` as a new config option of the container.

* Update doc and entrypoint

* Add link to kubernetes cron
* Remove erroneous `;` in the entrypoint

* Made build make target dependency of push